CLN move viz->plotting + improve cli flags
