+++ benchopt/plotting/__init__.py
