+++ benchopt/plotting/helpers.py
