+++ benchopt/plotting/plot_histogram.py
