+++ benchopt/plotting/plot_objective_curve.py
