+++ benchopt/plotting/plot_suboptimality_curve.py
